26 Improve calculations and lower render time
Prior to this change, frame rate was calculated off the total sum of
frames passed since program start. The text showing the current frame
rate updated every frame, causing it to hog the CPU and the GPU.

This change calculates frame rate via a moving average with a buffer of
100 frame times. The program now counts the amount of frames before
rendering the text, minimizing its resource load. Same is done for the
frame rate calculation.